+++ conformance/interfaces/kill/3-1.c
@@ -21,7 +21,10 @@
 
 int main()
 {
-        setuid(1); /* this is added incase user is root. If user is normal user, then it has no effect on the tests*/
+	if (getuid() == 0 && setuid(1) != 0) {
+		perror("setuid() failed");
+		return PTS_UNRESOLVED;
+	}
 
 	if (kill(1, 0) != -1) {
 		printf("Test FAILED: kill() succeeded even though this program's user id did not match the recieving process's user id\n");
